requirements.txt
Pillow
discord
git+https://github.com/Rapptz/discord.py@master
aiohttp
python-dateutil
asyncio
yarl<1.2
pytz
jishaku
luas.py
requests